--- a/frontend/device/generic/device.lua
+++ b/frontend/device/generic/device.lua
@@ -584,7 +584,7 @@ function Device:standby(max_duration) end
 
 -- Returns a string, used to determine the platform to fetch OTA updates
 function Device:otaModel()
-    return self.ota_model, "ota"
+    return self.ota_model, "kotasync"
 end
 
 --[[--