--- a/platform/pocketbook/koreader.app
+++ b/platform/pocketbook/koreader.app
@@ -41,8 +41,7 @@ echo $$ >/tmp/koreader.pid
 
 # update to new version from OTA directory
 ko_update_check() {
-    NEWUPDATE="${KOREADER_DIR}/ota/koreader.updated.tar"
-    INSTALLED="${KOREADER_DIR}/ota/koreader.installed.tar"
+    NEWUPDATE="${KOREADER_DIR}/ota/update.zip"
     if [ -f "${NEWUPDATE}" ]; then
         "${KOREADER_DIR}/fbink" -q -y -7 -pmh "Updating KOReader"
         # Keep a copy of the old manifest for cleaning leftovers later.
@@ -51,32 +50,11 @@ ko_update_check() {
         export FBINK_NAMED_PIPE="/tmp/.koreader.fbink"
         rm -f "${FBINK_NAMED_PIPE}"
         FBINK_PID="$("${KOREADER_DIR}/fbink" --daemon 1 %KOREADER% -q -y -6 -P 0)"
-        # NOTE: See frontend/ui/otamanager.lua for a few more details on how we squeeze a percentage out of tar's checkpoint feature
-        # NOTE: %B should always be 512 in our case, so let stat do part of the maths for us instead of using %s ;).
-        FILESIZE="$(stat -c %b "${NEWUPDATE}")"
-        # shellcheck disable=SC2003
-        BLOCKS="$(expr "${FILESIZE}" / 20)"
-        # shellcheck disable=SC2003
-        CPOINTS="$(expr "${BLOCKS}" / 100)"
-        export CPOINTS
-        # NOTE: We don't run as root, but folders created over USBMS are owned by root, which yields fun permission shenanigans...
-        #       c.f., https://github.com/koreader/koreader/issues/7581
-        KO_PB_TARLOG="/tmp/.koreader.tar"
-        # shellcheck disable=SC2016
-        "${KOREADER_DIR}/tar" --no-same-permissions --no-same-owner --checkpoint="${CPOINTS}" --checkpoint-action=exec='printf "%s" $(expr ${TAR_CHECKPOINT} / ${CPOINTS}) > ${FBINK_NAMED_PIPE}' -C "/mnt/ext1" -xf "${NEWUPDATE}" 2>"${KO_PB_TARLOG}"
+        (cd "${UNPACK_DIR}" && "${KOREADER_DIR}/unpack" -X "${NEWUPDATE}" >"${FBINK_NAMED_PIPE}")
         fail=$?
         kill -TERM "${FBINK_PID}"
-        # As mentioned above, filter out potential chmod & utime failures...
-        if [ "${fail}" -ne 0 ]; then
-            if [ "$(grep -Evc '(Cannot utime|Cannot change mode|Exiting with failure status due to previous errors)' "${KO_PB_TARLOG}")" -eq "0" ]; then
-                # No other errors, we're good!
-                fail=0
-            fi
-        fi
-        rm -f "${KO_PB_TARLOG}"
         # Cleanup behind us...
         if [ "${fail}" -eq 0 ]; then
-            mv "${NEWUPDATE}" "${INSTALLED}"
             # Cleanup leftovers from previous install.
             sed 's,^../,,' "${KOREADER_DIR}/ota/package.index" >/tmp/package.index.new
             (cd "${UNPACK_DIR}" && grep -xvFf /tmp/package.index.new /tmp/package.index.old | xargs -r rm -vf)
@@ -88,8 +66,7 @@ ko_update_check() {
             "${KOREADER_DIR}/fbink" -q -y -5 -pm "KOReader may fail to function properly!"
         fi
         rm -f /tmp/package.index.old /tmp/package.index.new "${NEWUPDATE}" # always purge newupdate to prevent update loops
-        unset CPOINTS FBINK_NAMED_PIPE
-        unset BLOCKS FILESIZE FBINK_PID
+        unset FBINK_NAMED_PIPE FBINK_PID
         # Ensure everything is flushed to disk before we restart. This *will* stall for a while on slow storage!
         sync
     fi