… frontend/ui/otamanager.lua
… platform/common/spinning_zsync
… platform/remarkable/spinning_zsync