# Current system based on zsync


The distribution is a TAR file (~105 MB) compressed with `gzip --rsyncable` (~43M). The zsync file is created from the tar.gz with `-C`, so the client end up with the decompressed tar file.

With the default block size of 2048, the resulting zsync manifest weights about 680 KB.

Consequently, fetching and installing an update uses a lot of disk space (~210MB) and even more IO.


### Disk IO / use

Taking for example the update from v2024.11 to v2025.04 (314 files changed, 64 MB):
- if `ota/koreader-installed.tar` is missing: 105 MB read + 105 MB written
- for zsync2:
    - saving the zsync file: 680 KB written
    - reading the seed file: 105 MB read
    - writing `ota/koreader-updated.tar`: 54 MB written from the seed, 85 MB written + 164 KB read while fetching the missing data
    - and to verify the file: 105 MB read, again
- then for extracting `ota/koreader-updated.tar` on restart: 105 MB read + 105 MB written

Total IO: 420 MB read + 350 MB written for 64 MB of changed files!

On my old kindle, disk IO is the limiting factor:
- time to extract the tar.gz with gnutar to /dev/null: 9.90s
- time to extract the tar.gz with gnutar to disk: 48.51s
- time to extract the tar with gnutar to /dev/null: 12.61s
- time to extract the tar with gnutar to disk: 67.31s


### Timings

On the kindle (with `ota/koreader-installed.tar` already present, running from command line):

| From ⭣ to ⭢ |  v2025.04 | v2025.04-90 | v2025.04-94 | v2025.04-95
|:-           |-:         |-:           |-:           |-:
| v2024.11    | 2m 24.16s |   2m 26.68s |   2m 33.44s |   2m 37.09s
| v2025.04    |           |   2m 33.02s |   2m 28.86s |   2m 16.53s
| v2025.04-90 |           |             |   1m 25.26s |   1m 27.58s       
| v2025.04-94 |           |             |             |   1m 15.09s

It takes almost a minute to process the seed file.

For a no change sync (`ota/koreader-installed.tar` matches 100%): 1m 15.04s.


### Network IO

Using #13888 as base (files repacked with `-mtime=… --numeric-owner`):

| From ⭣ to ⭢ |       v2025.04 |    v2025.04-90 |    v2025.04-94 |    v2025.04-95
|:-           |              -:|              -:|              -:|             -:
| v2024.11    | 2 KB ⮁ 30.6 MB | 2 KB ⮁ 30.2 MB | 2 KB ⮁ 30.2 MB | 2 KB ⮁ 30.2 MB
| v2025.04    |                | 2 KB ⮁ 25.0 MB | 2 KB ⮁ 25.0 MB | 2 KB ⮁ 25.0 MB
| v2025.04-90 |                |                | 2 KB ⮁  5.1 MB | 2 KB ⮁  5.6 MB
| v2025.04-94 |                |                |                | 1 KB ⮁  2.9 MB


### Memory use

Maximum memory use on desktop (resident set size of the process during its lifetime, as reported by time): ~12.2 MB.

Peak heap usage on desktop (as reported by massif): ~5 MB.


# New custom system (zipsync)


Refresher on the [ZIP file format](https://en.wikipedia.org/wiki/ZIP_(file_format)#File_headers):

```
 ┌────────────────────────┐
 │ LFH 1                  │
 │ (Local File Header)    │←╮
 ├────────────────────────┤ │
 │ compressed data        │ │
 ├────────────────────────┤ │
 │ LFH 2                  │ │←╮
 ├────────────────────────┤ │ │
 │ compressed data        │ │ │
 ├────────────────────────┤ │ │
 ┊                        ┊ │ │
 ┊                        ┊ │ │
 ├────────────────────────┤ │ │
 │ LFH n                  │ │ │←╮
 ├────────────────────────┤ │ │ │
 │ compressed data        │ │ │ │
 ├────────────────────────┤ │ │ │
 │ CDFH 1                 │╶╯ │ │
 │ (Central Directory FH) │   │ │
 ├────────────────────────┤   │ │
 │ CDFH 2                 │╶──╯ │
 ├────────────────────────┤     │
 ┊                        ┊     │
 ┊                        ┊     │
 ├────────────────────────┤     │
 │ CDFH n                 │╶────╯
 ├────────────────────────┤
 │ EOCD (End of CD)       │
 └────────────────────────┘
```

It's easy to reorder, or filter some entries to re-create a new ZIP, without needing to know the detail of the compression method(s), because each entry is compressed individually: just copy the LFH + compressed data for the files you want, and do the same for the central directory, then write the EOCD marker.

The base OTA distribution is now a ZIP file, compressed with XZ at maximum level, weighting about 35 MB (33 MB with binary gettext catalogs).

Decompressing the complete ZIP on the kindle takes 61.74s.

Similarly to zsync, a manifest is created: a JSON file ZSTD compressed.

Example for version `v2025.04`:
```json
{
  "filename": "koreader-kindlepw2-v2025.04.9.xz.7z.reordered.zip",
  "files": [
    {
      "hash": "05dbf7a5c1422d05",
      "path": "extensions/koreader/README.txt",
      "size": 636,
      "zip_hash": "4b5ab0f27b3b3714",
      "zip_start": 9222,
      "zip_stop": 9761
    },
    {
      "hash": "190f8327095a4640",
      "path": "extensions/koreader/bin/koreader-ext.sh",
      "size": 3021,
      "zip_hash": "670ba08bd108ce53",
      "zip_start": 9762,
      "zip_stop": 10994
    },
[…]
    {
      "hash": "54960240dcf8198a",
      "path": "koreader/tar",
      "size": 316228,
      "zip_hash": "1a47274934c76509",
      "zip_start": 34723517,
      "zip_stop": 34908594
    },
    {
      "hash": "b5d114e0103bc4a5",
      "path": "koreader/zsync2",
      "size": 721936,
      "zip_hash": "f46b97ab73e7c0ef",
      "zip_start": 34908595,
      "zip_stop": 35273483
    }
  ],
  "zip_cdir_hash": "cbca471ef8dca9b9",
  "zip_cdir_start": 35273484,
  "zip_cdir_stop": 35414028
}
```

Note: folders are ignored, permissions & times are not tracked.

For each file:
- `hash` is a 64bits hash hexadecimal digest (generated with [xxHash XXH3](https://github.com/Cyan4973/xxHash)) of the uncompressed data
- `size` is the size of the uncompressed data
- `zip_hash` is the hash of the LFH + compressed data ZIP block
- `zip_start`-`zip_end` are the start & end offsets of said block in the ZIP file

Other metadata:
- `filename`: self-explanatory
- `zip_cdir_hash`: hash of the whole central directory block (first CDFH to just before the EOCD)
- `zip_cdir_start-zip_cdir_stop`: offsets of said block

The resulting file weights about 37 KB.

On the platform, an OTA update is divided in 3 steps:
- fetching the manifest
- preparing the update (determine the list of changed / missing files): for each file in the manifest, add it to the changed / missing list if the path is not present locally, the size is different, or the hash does not match
- if they are changed / missing files, create the update (`ota/update.zip`):
  - fetch each file's corresponding ZIP block, checking its hash before writing it
  - fetch the whole central directory, check its hash, and filter it to write only the relevant entries
  - write the EOCD marker


### Disk IO / use

Disk IO for the same update from v2024.11 to v2025.04:
- for zipsync:
    - saving the zipsync file and associated etag: 46 KB written (the file is fetched in memory, decompressed and deserialized for use, then re-serialized / compressed with a lower compression level for saving, hence the size difference)
    - hash the files that are already present locally and have the same size: 40.7 MB read (689 files)
    - fetch and write the update file: 17.7 MB
- then for extracting `ota/update.zip` on restart: 17.7 MB read + 64 MB written

Total IO: 58.4 MB read + 82.2 MB written.

Disk usage: at most ~35 MB during the update, ~46 KB once the update is done.

More realistically: some files never change (e.g.: fonts), so except in case of corrupted install (missing and/or modified files), the upper bound for the update file size is closer to ~21 MB.


### Timings

On the kindle (running from the command line again):

| From ⭣ to ⭢ |  v2025.04 | v2025.04-90 | v2025.04-94 | v2025.04-95
|:-           |-:         |-:           |-:           |-:
| v2024.11    | 0m 24.41s |   0m 20.04s |   0m 27.09s |   0m 19.18s
| v2025.04    |           |   0m 26.21s |   0m 33.45s |   0m 30.36s
| v2025.04-90 |           |             |   0m 22.35s |   0m 15.78s
| v2025.04-94 |           |             |             |   0m 20.75s

No change sync: 0m 15.31s (basically the time it takes to hash the whole install).


### Network IO

| From ⭣ to ⭢ |         v2025.04 |      v2025.04-90 |      v2025.04-94 |      v2025.04-95
|:-           |                -:|                -:|                -:|               -:
| v2024.11    | 515 B  ⮁ 17.8 MB | 590 B  ⮁ 17.7 MB | 590 B  ⮁ 17.7 MB | 590 B  ⮁ 17.7 MB
| v2025.04    |                  | 590 B  ⮁ 17.2 MB | 590 B  ⮁ 17.2 MB | 590 B  ⮁ 17.2 MB
| v2025.04-90 |                  |                  | 590 B  ⮁  1.9 MB | 590 B  ⮁  2.0 MB
| v2025.04-94 |                  |                  |                  | 590 B  ⮁  333 KB

The update data is fetched using a number of HTTP range requests (`Range: bytes=start-stop`). Multi-range requests are not supported by the OTA server (because of Cloudflare's caching). Normally, this would result in lot more outgoing network traffic, even after merging adjacent ranges. Taking v2024.11 to v2025.04 as an example again: one request for zipsync + 181 range requests for the update = 33 KB ⮁ 17.9 MB. The trick to reduce that number is to process the distribution ZIP server' side, before generating the manifest, to reorder entries compared to the previous version:
```
# zipsync.py reorder old.zip new.zip
┌─────────────┬──────────────────┬────────────────────┬────┬──────┐
│   folders   │ unmodified files │ new/modified files │ CD │ EOCD │
│ (new order) │   (old order)    │    (new order)     │    │      │
└─────────────┴──────────────────┴────────────────────┴────┴──────┘
```

This ensure every forward (no local modifications) update can be done with only 2 requests: one to conditionally fetch the zipsync manifest, and another single range request to fetch the missing data, extending from somewhere in the unmodified files block (when lagging multiple versions behind), or the start of the new/modified files block (one step behind), to the end of the central directory.


### Memory

Maximum memory use on desktop (resident set size of the process during its lifetime, as reported by time): ~8.6 MB.

Peak heap usage on desktop (as reported by massif): ~1.3 MB.