@NiLuJe 2025-06-26
As another point of data, your timings are probably near worst-case (Kindle I/O sucks really bad), which means this might be even more impressive on devices with better storage.