@benoit-pierre 2025-07-13
Is xz available on all our zsync OTA platforms? It is on my kindle.