@NiLuJe 2025-07-13
> Is xz available on all our zsync OTA platforms? It is on my kindle.

I remember shipping it manually in FW packages, so that would be new-ish ;).

I'm also fairly sure it's not available on Kobo.

TL;DR: Probably not ;).