@NiLuJe 2025-07-13
> Can Windows users decompress a `tar.xz`?

Most unarchivers will do it in a two-step process (xz -> tar), which is weird and confusng, but, technically, yeah. (WinRAR used to do it in one step, but... WinRAR, lol).