@NiLuJe 2025-07-13
> What about macOS users?

I don't remember how Finder does it, but there's always Terminal anyway ;).