@benoit-pierre 2025-07-13
Anyway, on Kindle, that would be only for the KUAL side ("Update KOReader"), our unpack executable should already support it.

This could be an alternative to ZIP: it's possible to xz compress a TAR file with multiple blocks, one block per file. Example:
```bash
# Create the initial TAR:
▸ tar --create --no-recursion --numeric-owner --owner=0 --group=0 --verbatim-files-from --files-from=koreader/ota/package.index -f dist.tar 
# Get the start block for each file:
▸ tar tf dist.tar -R >tar.index
▸ tail -n5 tar.index
block 157031: koreader/unpack
block 157484: koreader/update_once.marker
block 157486: koreader/wmctrl
block 157538: koreader/zipsync.lua
block 157555: ** Block of NULs **
# Create list of blocks for xz:
▸ <tar.index awk -F '[: ]+' -v bs=512 -v total="$(stat -c %s dist.tar)" -v ORS=, 'BEGIN { total = total / bs } { if ($2) print ($2-prev)*bs; prev=$2 } END { print (total-prev)*512 }' >blocklist
▸ dd if=blocklist count=1 status=none bs=64; echo
512,512,1536,512,3584,3072,1024,2560,512,35328,7168,512,10240,51
# Compress the TAR:
▸ xz -9 --block-list="$(cat blocklist)" dist.tar
▸ xz -lv dist.tar.xz | head -n20
dist.tar.xz (1/1)
  Streams:           1
  Blocks:            1,173
  Compressed size:   30.0 MiB (31,508,036 B)
  Uncompressed size: 76.9 MiB (80,670,720 B)
  Ratio:             0.391
  Check:             CRC64
  Stream Padding:    0 B
  Streams:
    Stream    Blocks      CompOffset    UncompOffset        CompSize      UncompSize  Ratio  Check      Padding
         1     1,173               0               0      31,508,036      80,670,720  0.391  CRC64            0
  Blocks:
    Stream     Block      CompOffset    UncompOffset       TotalSize      UncompSize  Ratio  Check
         1         1              12               0              96             512  0.188  CRC64
         1         2             108             512             108             512  0.211  CRC64
         1         3             216           1,024             520           1,536  0.339  CRC64
         1         4             736           2,560             112             512  0.219  CRC64
         1         5             848           3,072           1,200           3,584  0.335  CRC64
         1         6           2,048           6,656           1,248           3,072  0.406  CRC64
         1         7           3,296           9,728             280           1,024  0.273  CRC64
```

Then use the same mechanism as above, but with an XZ file.