@NiLuJe 2025-07-13
> I'm on 5.12.2.2, so not that fresh, no?

I stopped caring around 5.8 or so, so, still fairly new to me ;p.