@NiLuJe 2025-07-13
> The 2015 firmware I downloaded does have it. On my kindle it's part of busybox, so maybe not that uncommon.

It indeed seems to be in there, my bad ;).

(I ship a trimmed down xzdec so I rarely check for the standard names ;p)