@benoit-pierre 2025-08-31
OK, so I ended up rewriting the code to use `tar.xz` as the format: even if the ZIP standard does officially support a bunch of different compression algorithms, the standard tools (i.e. the unzip executable available on Linux systems, or official support on Windows) don't support anything else except "deflate" (and strangely bzip2 on my Arch Linux system, go figure…). This breaks expectations: a ZIP that can't be decompressed. 

Switching to `tar.xz` as several advantages:
1. supported on Linux (including (my) Kindle, Kobo, and Pocketbook), macOS, and Windows (>= 11)
2. we can use the xz `liblzma` library API to handle files
3. which means we can easily recreate the stream index (the equivalent of a ZIP's central directory block) and there's no need to fetch it
4. consequently, reordering now move changed files at the beginning

Thanks to 3 & 4, Network IO is slightly reduced:
| From ⭣ to ⭢ |        v2025.04 |     v2025.04-90 |     v2025.04-94 |     v2025.04-95
|:-           |               -:|               -:|               -:|              -:
| v2024.11    | 343 B ⮁ 17.7 MB | 393 B ⮁ 17.6 MB | 393 B ⮁ 17.6 MB | 393 B ⮁ 17.6 MB
| v2025.04    |                 | 393 B ⮁ 17.1 MB | 393 B ⮁ 17.1 MB | 393 B ⮁ 17.1 MB
| v2025.04-90 |                 |                 | 392 B ⮁  1.7 MB | 392 B ⮁  1.9 MB
| v2025.04-94 |                 |                 |                 | 391 B ⮁  191 KB


The manifest format is similar:
```json
{
  "filename": "koreader-kindlepw2-v2025.04.tar.xz",
  "files": [
    {
      "hash": "85512b33e607c462",
      "path": "koreader/common/Spore.lua",
      "size": 9537,
      "xz_hash": "0f25a8637fde08be",
      "xz_offset": 12,
      "xz_size": 2666
    },
[…]
    {
      "hash": "75eafcd0f43d0efa",
      "path": "koreader/fonts/urw/Dingbats.cff",
      "size": 29728,
      "xz_hash": "98151aced2048a44",
      "xz_offset": 35249656,
      "xz_size": 24065
    }
  ],
  "xz_check": 4
}
```

Said manifest can now be created from `koreader/ota/package.index`, e.g.: `kotasync --reorder previous_version.kotasync --manifest koreader/ota/package.index new_version.tar.xz`. This way there's no need create a both a ZIP and a `tar.xz`: the later can contain everything, and an OTA update will not touch `system/bin/koreader.app` on Pocketbook, or `koreader.png` on Kobo.