@benoit-pierre 2025-08-31
Merge plan:
- [ ] https://github.com/koreader/crengine/pull/623
- [ ] https://github.com/koreader/koreader-base/pull/2113
- [ ] https://github.com/koreader/koreader/pull/14182
- [ ] this PR
- [ ] https://github.com/koreader/virdevenv/pull/123