@benoit-pierre 2025-08-31
Minor issue: there are some weird UI refresh states with respected to progress dialogs.


https://github.com/user-attachments/assets/837d7ae9-9133-4896-8199-5a74ad960b28