@Frenzie 2025-08-31
> This way there's no need create a both a ZIP and a tar.xz

You mean specifically for OTA purposes? I suspect the average user would still like a zip.