@benoit-pierre 2025-08-31
I was thinking both, so there's no need to create 2 different archives.