@Frenzie 2025-08-31
I can't get behind such an additional obstacle, although it would be nice to have technically. I'll always want the program to be a potential entry point to some simple hacking in a sense that may have never happened by first having to install 7-zip, no matter how simple that may seem. Such small inconveniences can have a large impact. :-)