@NiLuJe 2025-08-31
> Minor issue: there are some weird UI refresh states with respected to progress dialogs.
> kotasync.mp4

That's... weird :D. I'm assuming it's simply the initial onShow refresh that gets a cropped refresh region for some reason?