@mergen3107 2026-08-18
@benoit-pierre 

Finally tried OTA update!
It is blazing fast and efficient! Thank you :D 

Sorry for 360p video quality 

https://github.com/user-attachments/assets/a55fae8c-eeb9-43e3-881a-9a9dc1cc5d97